feat: Create CustomHook
1. Create useCardEvent to handle focus & hover event
